不具合修正

+++ Project-HEW2020/overlay.cpp
@@ -29,6 +29,6 @@ void GameOverlay::LoadTexture(const char* pFileName)
 	pSprite = new SpriteNormal;
 	pSprite->LoadTexture(pFileName);
 
-	width = pSprite->GetTextureWidth();
-	height = pSprite->GetTextureHeight();
+	width = (float)pSprite->GetTextureWidth();
+	height = (float)pSprite->GetTextureHeight();
 }
